requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
tiktoken>=0.7.0
